Refining validation approach
Validation is now based on split of image directories not a split of image
This means that the viewer can show validation over a video stream and not
random images.

Modified generator to handle this.
Modified viewer for new generator
Added ability to save the validation image list dataframe to a file for use in the viewer